Move tests into src directory
